chat preview added
